FEATURE: 자산의 잔액 합계 조회 시 예외 처리 추가 FIX: 잔액 합계를 int 타입에서 Long 타입으로 수정

+++ src/test/java/com/iiiiii/accountbook/asset/query/service/AssetServiceTests.java
@@ -32,15 +32,15 @@ public class AssetServiceTests {
     @ParameterizedTest
     @ValueSource(ints = 2)
     public void findMyAllAssetsAndBalance(int memberCode) {
-        int totalBalance = assetService.calcTotalBalance(memberCode);
-        assertEquals(6000000, totalBalance);
+        Long totalBalance = assetService.calcTotalBalance(memberCode);
+        assertEquals(6000000L, totalBalance);
     }
 
     @DisplayName("한 회원의 카테고리별 자산과 잔액 합계 조회 테스트")
     @ParameterizedTest
     @MethodSource("provideMemberCodeAndCategory")
     public void findMyAssetsAndBalance(int memberCode, AssetCategory category) {
-        int totalBalance = assetService.calcTotalBalanceByCategory(memberCode, category);
-        assertEquals(25000000, totalBalance);
+        Long totalBalance = assetService.calcTotalBalanceByCategory(memberCode, category);
+        assertEquals(25000000L, totalBalance);
     }
 }
